+++ packages/create-cloudflare/src/templates.ts
@@ -244,6 +244,10 @@ export function getFrameworkMap({ experimental = false }): TemplateMap {
 			"tanstack-start": tanStackStartTemplate,
 			angular: angularTemplate,
 			solid: solidTemplate,
+			qwik: qwikTemplate,
+			vue: vueTemplate,
+			react: reactTemplate,
+			"react-router": reactRouterTemplate,
 		};
 	} else {
 		return {