+++ packages/wrangler/src/autoconfig/frameworks/get-framework.ts
@@ -1,5 +1,7 @@
 import { Angular } from "./angular";
 import { Astro } from "./astro";
+import { Qwik } from "./qwik";
+import { ReactRouter } from "./react-router";
 import { SolidStart } from "./solid-start";
 import { Static } from "./static";
 import { SvelteKit } from "./sveltekit";
@@ -17,10 +19,14 @@ export function getFramework(detectedFramework?: {
 			return new SvelteKit(detectedFramework.name);
 		case "tanstack-start":
 			return new TanstackStart(detectedFramework.name);
+		case "react-router":
+			return new ReactRouter(detectedFramework.name);
 		case "angular":
 			return new Angular(detectedFramework.name);
 		case "solid-start":
 			return new SolidStart(detectedFramework.name);
+		case "qwik":
+			return new Qwik(detectedFramework.name);
 		default:
 			return new Static(detectedFramework?.name);
 	}